api/requirements.txt
fastapi==0.111
langchain_chroma==0.1.2
langchain_community==0.2.10
langchain_core==0.2.24
langchain_openai==0.1.19
langchain_text_splitters==0.2.2
pydantic==1.10.8
PyJWT==2.8.0
pymongo==3.12.1
python-dotenv==1.0.1
tiktoken==0.7.0
stripe>=10.0.0
chromadb==0.5.0
uvicorn[standard]==0.23.2
pyjwt[crypto]==2.8.0
pysqlite3-binary